+++ src/app/lib/data/TeamData.js
@@ -1,190 +1,10 @@
-export const TeamData2k21 =[
-    {
-        name:'Akash kumar',
-        pos:'President',
-        imgSrc:'akash_kumar.png',
-        socials:{
-            github:'',
-            email:'akashashokasha@gmail.com',
-            linkedin:'akash-kumar-7b4086245',
-            insta:'a.k.ash.k.um.ar',
-        },
-    },
-    {
-        name:'Samia Nishaat',
-        pos:'Vice President',
-        imgSrc:'samia_nishaat.png',
-        socials:{
-            github:'',
-            email:'samiabksc@gmail.com',
-            linkedin:'samianishaat',
-            insta:'samianishaat',
-        },
-    },
-    
-    {
-        name:'Priyanshu katyayni',
-        pos:'Vice President',
-        imgSrc:'priyanshu_katyayni.png',
-        socials:{
-            github:'lonewolf2299',
-            email:'priyanshukatyayni229@gmail.com',
-            linkedin:'priyanshu-katyayni-70b0a4233',
-            insta:'priyanshu._.13',
-        },
-    },
-    
-    {
-        name:'Rohit Kumar',
-        pos:'General Secretary',
-        imgSrc:'rohit_kumar.png',
-        socials:{
-            github:'rohitkumar9897',
-            email:'rohitku8210610477@gmail.com',
-            linkedin:'rohitkumar--/',
-            insta:'_rohit_k.u.m.a.r_',
-        },
-    },
-    {
-        name:'Mohit Pandey ',
-        pos:'Technical Secretary',
-        imgSrc:'mohit_pandey.png',
-        socials:{
-            github:'MOHIT-PANDEY215',
-            email:'pandeymohit215@gmail.com',
-            linkedin:'mohit-pandey-5b0756197',
-            insta:'pandeymohit215',
-        },
-    },
-    {
-        name:'Aditya Bhardwaj',
-        pos:'Technical Head',
-        imgSrc:'aditya_bhardwaj.png',
-        socials:{
-            github:'adityaS49',
-            email:'aditya.s21@outlook.com',
-            linkedin:'aditya-bhardwaj-01027b21b/',
-            insta:'_aditya.s21__',
-        },
-    },   
-    {
-        name:'Tushar kumar',
-        pos:'Joint-Secretary & Pro',
-        imgSrc:'tushar_kumar.png',
-        socials:{
-            github:'',
-            email:'tusharkumar121002@gmail.com',
-            linkedin:'',
-            insta:'Tushar_1729',
-        },
-    },
-    {
-        name:'Esha Kumari',
-        pos:'Joint Secretary',
-        imgSrc:'esha_kumari.png',
-        socials:{
-            github:'Esha-01',
-            email:'eshakumari2002@gmail.com',
-            linkedin:'esha-kumari-47b854232',
-            insta:'wassup_esha',
-        },
-    },
-    {
-        name:'Dipanshu Ranjan',
-        pos:'Treasurer',
-        imgSrc:'dipanshu_ranjan.png',
-        socials:{
-            github:'',
-            email:'dtdtstst1@gmail.com',
-            linkedin:'',
-            insta:'Dipanshu1133',
-        },
-    },
-    {
-        name:'Aayushi',
-        pos:'Treasurer & Alumni Incharge',
-        imgSrc:'aayushi_arya.png',
-        socials:{
-            github:'Aarya1527',
-            email:'aayushi.khu@gmail.com',
-            linkedin:'aayushi-aarya-820138231/',
-            insta:'aayushi_aarya15',
-        },
-    },
-    {
-        name:'Abhay  Narayan',
-        pos:'Joint-Treasurer & Event Coordinator',
-        imgSrc:'abhay_narayan.png',
-        socials:{
-            github:'naraynabhay',
-            email:'abhaynarayan07062002@gmail.com',
-            linkedin:'abhay-narayan-76114b1ba',
-            insta:'abhay_narayan_00',
-        },
-    },
-    {
-        name:'Rahul kumar',
-        pos:'Event Coordinator',
-        imgSrc:'rahul_kumar.png',
-        socials:{
-            github:'',
-            email:'01rahulkumaar@gmail.com',
-            linkedin:'rahul-kumar-482219238',
-            insta:'rahul.kumar11',
-        },
-    },
-    {
-        name:'Aayushi Singh',
-        pos:'Head Editor',
-        imgSrc:'aayushi_singh.png',
-        socials:{
-            github:'',
-            email:'aayushisingh086@gmail.com',
-            linkedin:'aayushi-singh-5107bb256',
-            insta:'',
-        },
-    },
-    
-    {
-        name:'Chinmay Gari',
-        pos:'Alumni Incharge',
-        imgSrc:'chinmay_gari.png',
-        socials:{
-            github:'',
-            email:'chinmaygari@gmail.com',
-            linkedin:'',
-            insta:'__ch_in_may__',
-        },
-    },
-    {
-        name:'Kheya Rani Das',
-        pos:'Public Relation Officer',
-        imgSrc:'kheya_das.png',
-        socials:{
-            github:'kheya19',
-            email:'kheyadas1919@gmail.com',
-            linkedin:'kheya-das-73266921a',
-            insta:'kheya__das',
-        },
-    },
-         
-]
+
 
 export const TeamData2k22 =[
-    {
-        name:'Amaan Hussain',
-        pos:'',
-        imgSrc:'Amaan2.png',
-        socials:{
-            github:'Amaanhussain11',
-            email:'amaanhussain983@gmail.com',
-            linkedin:'amaanhussain01',
-            insta:'amaan_hussain_011',
-        }
-    },
+    
     {
         name:'Shobhit Kashyap',
-        pos:'',
+        pos:'President',
         imgSrc:'shobhit_kashyap1.png',
         socials:{
             github:'imshobhitk',
@@ -194,19 +14,52 @@ export const TeamData2k22 =[
         },
     },
     {
-        name:'Gourav Mondal',
-        pos:'',
-        imgSrc:'Gourav2.png',
+        name:'Shreya Paul',
+        pos:'Vice President',
+        imgSrc:'shreya_paul.png',
         socials:{
-            github:'Gourav123-sys',
-            email:'mondalgourav227@gmail.com',
-            linkedin:'gourav-mondal-55a812204',
-            insta:'gour_av_30',
+            github:'shreyapaul01',
+            email:'shreyapaul100@yahoo.com',
+            linkedin:'shreya-paul-a84606257',
+            insta:'its.me._.shreya',
+        },
+    },
+    {
+        name:'Deepak kumar Rajak',
+        pos:'Vice President',
+        imgSrc:'deepak_rajak.png',
+        socials:{
+            github:'Iamdeepak4551 ',
+            email:'iamdeepak4551@gmail.com',
+            linkedin:'deepak-kumar-rajak-41a9bb222',
+            insta:'Iamdeepak4551',
+        },
+    },
+    {
+        name:'Harsh Raj',
+        pos:'Secretary',
+        imgSrc:'harsh.png',
+        socials:{
+            github:'',
+            email:'rharsh808@gmail.com',
+            linkedin:'itsharshraj',
+            insta:'tech_nicalharsh',
+        }
+    },
+    {
+        name:'Amaan Hussain',
+        pos:'Technical Secretary',
+        imgSrc:'Amaan2.png',
+        socials:{
+            github:'Amaanhussain11',
+            email:'amaanhussain983@gmail.com',
+            linkedin:'amaanhussain01',
+            insta:'amaan_hussain_011',
         }
     },
     {
         name:'Om Kumar Soni',
-        pos:'',
+        pos:'Techinal Head',
         imgSrc:'om_soni2.png',
         socials:{
             github:'omkumarsoni2919',
@@ -217,7 +70,7 @@ export const TeamData2k22 =[
     },
     {
         name:'Ashish Kumar Mahto',
-        pos:'',
+        pos:'Joint Secretary',
         imgSrc:'ashish_kumar.jpg',
         socials:{
             github:'ashish3180',
@@ -228,7 +81,7 @@ export const TeamData2k22 =[
     },
     {
         name:'Amisha Kumari Soren ',
-        pos:'',
+        pos:'Joint Secretary',
         imgSrc:'amisha_kumari.png',
         socials:{
             github:'',
@@ -237,61 +90,10 @@ export const TeamData2k22 =[
             insta:'amisha_soren8',
         },
     },
-    {
-        name:'Shreya Paul',
-        pos:'',
-        imgSrc:'shreya_paul.png',
-        socials:{
-            github:'shreyapaul01',
-            email:'shreyapaul100@yahoo.com',
-            linkedin:'shreya-paul-a84606257',
-            insta:'its.me._.shreya',
-        },
-    },
-    {
-        name:'Deepak Sawaiyan',
-        pos:'',
-        imgSrc:'deepak_sewaiyan.png',
-        socials:{
-            github:'Deepak200385',
-            email:'dk4142918@gmail.com',
-            linkedin:'deepak-sawaiyan-725975226	Deepak200385',
-            insta:'dee_pak_sawaiyan',
-        },
-    },
-    {
-        name:'Abhay Kumar',
-        pos:'',
-        imgSrc:'abhay_kumar.png',
-        socials:{
-            github:'leanwithkumar',
-            email:'mrhustle936@gmail.com',
-            linkedin:'abhay-kumar-443b981b6',
-            insta:'leanwithkumar',
-        }
-    },
-    
-    
-    
-    
-   
-    {
-        name:'Deepak kumar Rajak',
-        pos:'',
-        imgSrc:'deepak_rajak.png',
-        socials:{
-            github:'Iamdeepak4551 ',
-            email:'iamdeepak4551@gmail.com',
-            linkedin:'deepak-kumar-rajak-41a9bb222',
-            insta:'Iamdeepak4551',
-        },
-    },
-   
-    
     
     {
         name:'Manish Kumar',
-        pos:'',
+        pos:'Treasurer & Graphics Head',
         imgSrc:'manish.png',
         socials:{
             github:'',
@@ -301,16 +103,91 @@ export const TeamData2k22 =[
         }
     },
     {
-        name:'Harsh Raj',
-        pos:'',
-        imgSrc:'harsh.png',
+        name:'Gourav Mondal',
+        pos:'Joint Treasurer',
+        imgSrc:'Gourav2.png',
+        socials:{
+            github:'Gourav123-sys',
+            email:'mondalgourav227@gmail.com',
+            linkedin:'gourav-mondal-55a812204',
+            insta:'gour_av_30',
+        }
+    },
+    {
+        name:'Shivam Raj',
+        pos:'Joint Treasurer & Event Coordinator',
+        imgSrc:'shivam_raj.png',
         socials:{
             github:'',
-            email:'rharsh808@gmail.com',
-            linkedin:'itsharshraj',
-            insta:'tech_nicalharsh',
+            email:'raj.grad03@gmail.com',
+            linkedin:'',
+            insta:'',
+        },
+    },
+    {
+        name:'Shikha Rani Singh',
+        pos:'Editor in Chief',
+        imgSrc:'shikha_rani.png',
+        socials:{
+            github:'',
+            email:'shikharanisingh639@gmail.com',
+            linkedin:'shikha-singh-11824a2a3',
+            insta:'',
+        },
+    },
+
+    {
+        name:'Chahat Singh',
+        pos:'Alumni in Charge',
+        imgSrc:'chahat_singh.png',
+        socials:{
+            github:'',
+            email:'',
+            linkedin:'',
+            insta:'',
+        },
+    },
+    
+    
+    
+    
+    {
+        name:'Deepak Sawaiyan',
+        pos:'Design Head',
+        imgSrc:'deepak_sewaiyan.png',
+        socials:{
+            github:'Deepak200385',
+            email:'dk4142918@gmail.com',
+            linkedin:'deepak-sawaiyan-725975226	Deepak200385',
+            insta:'dee_pak_sawaiyan',
+        },
+    },
+    
+    
+    {
+        name:'Abhay Kumar',
+        pos:'Treasurer',
+        imgSrc:'abhay_kumar.png',
+        socials:{
+            github:'leanwithkumar',
+            email:'mrhustle936@gmail.com',
+            linkedin:'abhay-kumar-443b981b6',
+            insta:'leanwithkumar',
         }
     }
+
+    
+    
+    
+    
+    
+   
+    
+   
+    
+    
+    
+    
          
 ]
